Se cambia el tipo de dato de las columnas de precio de los materiales requeridos

+++ database/migrations/2015_09_10_184814_create_materiales_requeridos_table.php
@@ -16,11 +16,11 @@ class CreateMaterialesRequeridosTable extends Migration
             $table->increments('id');
             $table->integer('id_tipo_area')->unsigned()->index();
             $table->integer('id_material')->unsigned()->index();
-            $table->integer('cantidad_requerida')->default(1);
+            $table->decimal('cantidad_requerida', 9, 2)->default(1);
             $table->integer('id_moneda')->unsigned()->nullable()->index();
             $table->integer('precio_estimado')->default(0)->nullable();
             $table->boolean('se_evalua')->default(false);
-            $table->integer('cantidad_comparativa')->nullable();
+            $table->decimal('cantidad_comparativa', 9, 2)->nullable();
             $table->integer('id_moneda_comparativa')->unsigned()->nullable()->index();
             $table->integer('precio_comparativa')->nullable();
             $table->boolean('existe_para_comparativa')->default(true);
